refs #7602
Modified LogAggregationScheduleManager removed the pause for due to hazelcast (hazelcast is no longer a factor in job scheduling) before triggering jobs. 

Modified LogAggregationScheduleManager to shortcircuit rescheduling if there are no jobs to be scheduled or deleted.

Modified LogAggregationScheduleManager to only consider nodes whose state is UP when performing collection math to determine jobs to schedule or delete.